Rename node entry file. Rename events to events-node.
