+++ test/events.test.js
@@ -1,6 +1,6 @@
 const expect = require('chai').expect
 
-const getEvents = require('../src/plugins/events').default
+const getEvents = require('../src/plugins/events-node').default
 
 describe('events', () => {
   let events
